+++ CHANGELOG.md
@@ -31,6 +31,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Services directory
+- Category model
 
 ## [0.1.0] - 2025-06-16 - Vlad